requirements.txt
python-dotenv==0.20.0
requests==2.28.0
psycopg2==2.9.3
pprintpp==0.4.0
inquirer==2.9.2